--- src/main.ts
@@ -1,2 +1,0 @@
-export { isMaster } from "./lib/isMaster";
-export { getAction } from "./lib/getAction";
